Moves FormFieldRaw
